Refactor auth backends and update configuration
