+++ tests/unit/config.rs
@@ -26,7 +26,7 @@ max_article_bytes = "20K"
 fn runtime_update_preserves_immutable_fields() {
     let initial = r#"addr = ":119"
 db_path = "/tmp/db1"
-auth_db_path = "/tmp/auth1"
+auth_db_path = "sqlite:///tmp/auth1"
 peer_db_path = "/tmp/peer1"
 peer_sync_secs = 1800
 tls_addr = ":563"
@@ -42,7 +42,7 @@ retention_days = 5
 
     let updated = r#"addr = ":42"
 db_path = "/tmp/db2"
-auth_db_path = "/tmp/auth2"
+auth_db_path = "sqlite:///tmp/auth2"
 peer_db_path = "/tmp/peer2"
 peer_sync_secs = 3600
 tls_addr = ":9999"
@@ -59,7 +59,7 @@ retention_days = 1
 
     assert_eq!(cfg.addr, ":119");
     assert_eq!(cfg.db_path, "/tmp/db1");
-    assert_eq!(cfg.auth_db_path, "/tmp/auth1");
+    assert_eq!(cfg.auth_db_path, "sqlite:///tmp/auth1");
     assert_eq!(cfg.peer_db_path, "/tmp/peer1");
     assert_eq!(cfg.peer_sync_secs, 3600);
     assert_eq!(cfg.tls_addr.as_deref(), Some(":563"));
